* 'develop' of https://gitlab.com/jusfoun-bigdata/hookah:
  站内信
  实名认证 zc
  zc
  优化
  供应商管理完成
  zc
  bug
  邮件下发地址更换及代码完善
  fix
  fix bug
  结算处理
  邮件下发地址更换及代码完善

+++ hookah-core/src/main/java/com/jusfoun/hookah/core/domain/bo/MoneyInOutBo.java
@@ -8,7 +8,7 @@ public class MoneyInOutBo {
 
     public Long Money;          //  默认已经乘以100了
 
-    public byte operatorType;
+    public Integer operatorType;
 
     public String getUserId() {
         return userId;
@@ -34,11 +34,11 @@ public class MoneyInOutBo {
         Money = money;
     }
 
-    public byte getOperatorType() {
+    public Integer getOperatorType() {
         return operatorType;
     }
 
-    public void setOperatorType(byte operatorType) {
+    public void setOperatorType(Integer operatorType) {
         this.operatorType = operatorType;
     }
 }
